+++ packages/local-webrtc-browser-sdk/src/sdk/guest-local-webrtc-sdk.ts
@@ -1,14 +1,16 @@
 import { ArmdozerId, PilotId } from "gbraver-burst-core";
+import { nanoid } from "nanoid";
 import { Observable } from "rxjs";
 
+import { gatherAllIceCandidates } from "../webrtc/gather-all-ice-candidate";
 import { sendGuestMessage } from "../webrtc/guest/guest-message";
 import { receiveBattleStart } from "../webrtc/guest/receive-battle-start";
 import { receiveRequestSelectedPlayer } from "../webrtc/guest/receive-request-selected-player";
 import { waitUntilConnected } from "../webrtc/wait-until-connected";
-import { waitUntilIceCandidate } from "../webrtc/wait-untilIce-candidate";
-import { joinRoom } from "../ws-signal/join-room";
-import { sendGuestSignal } from "../ws-signal/send-guest-signal";
+import { joinRoom } from "../websocket/join-room";
+import { sendGuestSignal } from "../websocket/send-guest-signal";
 import { BattleSDK } from "./battle-sdk";
+import { FrontendLogManager } from "./frontend-log-manager";
 import { GuestBattleSDK } from "./guest-battle-sdk";
 import {
   GuestWebRTCConnectionManager,
@@ -58,6 +60,8 @@ export type GuestLocalWebRTCSDK = {
 type LocalWebRTCGuestSDKImplOptions = GuestWebRTCConnectionManagerOptions & {
   /** WebSocketシグナルサーバーのURL */
   wsSignalUrl: string;
+  /** WebRTCヘルパーAPIのURL */
+  webRTCHelperApiURL: string;
 };
 
 /** ローカルWebRTCゲスト用SDKの実装 */
@@ -66,6 +70,8 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
   #webRTCConnection: GuestWebRTCConnectionManager;
   /** WebSocketコネクションマネージャー */
   #websocketConnection: WebSocketConnectionManager;
+  /** フロントエンドログマネージャー */
+  #frontendLog: FrontendLogManager;
 
   /**
    * コンストラクタ
@@ -74,6 +80,7 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
   constructor(options: LocalWebRTCGuestSDKImplOptions) {
     this.#webRTCConnection = new GuestWebRTCConnectionManager(options);
     this.#websocketConnection = new WebSocketConnectionManager(options);
+    this.#frontendLog = new FrontendLogManager(options);
   }
 
   /** @override */
@@ -82,6 +89,7 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
     armdozerId: ArmdozerId;
     pilotId: PilotId;
   }) {
+    const spanId = nanoid();
     const { roomID, armdozerId, pilotId } = options;
 
     this.#websocketConnection.gracefulDisconnect();
@@ -98,7 +106,7 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
       return await receiveBattleStart(dataChannel);
     })();
 
-    const isSignalingSuccessful = await this.#signaling(roomID);
+    const isSignalingSuccessful = await this.#signaling({ roomID, spanId });
     if (!isSignalingSuccessful) {
       return null;
     }
@@ -139,10 +147,16 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
 
   /**
    * シグナリングを行う
-   * @param roomID ルームID
-   * @returns シグナリングが完了したら発火するPromise
+   * @param options シグナリングのオプション
+   * @param options.roomID ルームID
+   * @param options.spanId ログ用の識別子
+   * @returns シグナリングが完了したらtrue、失敗したらfalse
    */
-  async #signaling(roomID: string) {
+  async #signaling(options: {
+    roomID: string;
+    spanId: string;
+  }): Promise<boolean> {
+    const { roomID, spanId } = options;
     try {
       const websocket = await this.#websocketConnection.getOrCreate();
       const joinRoomAccepted = await joinRoom({ websocket, roomID });
@@ -159,11 +173,16 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
         hostIceCandidates.map((c) => connection.addIceCandidate(c)),
       );
       const guestSDP = await connection.createAnswer();
+
+      await this.#frontendLog.log({ type: "ICE_CANDIDATE_START", spanId });
       const [guestIceCandidates] = await Promise.all([
         // icecandidateイベントはsetLocalDescriptionの後に発生するため、先に待機しておく
-        waitUntilIceCandidate(connection),
+        gatherAllIceCandidates(connection),
         connection.setLocalDescription(guestSDP),
       ]);
+      await this.#frontendLog.log({ type: "ICE_CANDIDATE_END", spanId });
+
+      await this.#frontendLog.log({ type: "SIGNALING_START", spanId });
       const { reservationID } = joinRoomAccepted;
       await Promise.all([
         sendGuestSignal({
@@ -171,10 +190,14 @@ class GuestLocalWebRTCSDKImpl implements GuestLocalWebRTCSDK {
           roomID,
           reservationID,
           sdp: guestSDP,
-          iceCandidates: guestIceCandidates,
+          iceCandidates: guestIceCandidates.iceCandidates,
         }),
         waitUntilConnected(connection),
+        ...guestIceCandidates.iceCandidateErrors.map((error) =>
+          this.#frontendLog.log({ type: "ICE_CANDIDATE_ERROR", spanId, error }),
+        ),
       ]);
+      await this.#frontendLog.log({ type: "SIGNALING_END", spanId });
       return true;
     } finally {
       this.#websocketConnection.gracefulDisconnect();