+++ packages/local-webrtc-stub/src/use-case/guest-player.ts
@@ -1,6 +1,7 @@
 import { GuestLocalWebRTCSDK } from "@gbraver-burst-network/local-webrtc-browser-sdk";
 import { ArmdozerIds, PilotIds } from "gbraver-burst-core";
 
+import { waitTime } from "../wait-time";
 import { UseCase, UseCaseContext } from "./use-case";
 
 /** ゲスト側プレイヤー */
@@ -44,27 +45,33 @@ export class GuestPlayer implements UseCase {
       battery: 3,
     });
     console.log(update01);
+    await waitTime(1000);
     const update02 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 5,
     });
     console.log(update02);
+    await waitTime(1000);
     const update03 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 5,
     });
     console.log(update03);
+    await waitTime(1000);
     const update04 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 0,
     });
     console.log(update04);
+    await waitTime(1000);
     const update05 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 0,
     });
     console.log(update05);
+    await waitTime(1000);
 
+    console.log("end battle");
     this.#guestSDK.disconnectWebRTC();
     this.#guestSDK.disconnectWebSocket();
   }