+++ packages/local-webrtc-stub/src/use-case/host-player.ts
@@ -1,6 +1,7 @@
 import { HostLocalWebRTCSDK } from "@gbraver-burst-network/local-webrtc-browser-sdk";
 import { ArmdozerIds, PilotIds } from "gbraver-burst-core";
 
+import { waitTime } from "../wait-time";
 import { UseCase } from "./use-case";
 
 /** ホスト側プレイヤー */
@@ -35,31 +36,40 @@ export class HostPlayer implements UseCase {
 
     console.log("create room success, roomID:", room.roomID);
     const battle = await room.waitUntilMatching();
+    battle.suddenlyBattleEndNotifier().subscribe(() => {
+      console.log("suddenly battle end");
+    });
     console.log("start battle", battle);
     const update01 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 5,
     });
     console.log(update01);
+    await waitTime(1000);
     const update02 = await battle.progress({
       type: "BURST_COMMAND",
     });
     console.log(update02);
+    await waitTime(1000);
     const update03 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 5,
     });
     console.log(update03);
+    await waitTime(1000);
     const update04 = await battle.progress({
       type: "PILOT_SKILL_COMMAND",
     });
     console.log(update04);
+    await waitTime(1000);
     const update05 = await battle.progress({
       type: "BATTERY_COMMAND",
       battery: 5,
     });
     console.log(update05);
+    await waitTime(1000);
 
+    console.log("end battle");
     this.#hostSDK.disconnectWebRTC();
     this.#hostSDK.disconnectWebSocket();
   }